+++ include/gputils/pybind11_utils.hpp
@@ -22,7 +22,7 @@ namespace gputils {
 // for tracing/debugging array conversion.
 
 extern void convert_array_from_python(
-    void *&data, int &ndim, ssize_t *shape, ssize_t *strides, ssize_t &size,
+    void *&data, int &ndim, long *shape, long *strides, long &size,
     int dlpack_type_code, int itemsize, std::shared_ptr<void> &base, int &aflags,
     PyObject *src, bool convert, const char *debug_prefix = nullptr);
 
@@ -31,7 +31,7 @@ extern void convert_array_from_python(
 // On failure, calls PyErr_SetString() and returns NULL.
 
 extern PyObject *convert_array_to_python(
-    void *data, int ndim, const ssize_t *shape, const ssize_t *strides,
+    void *data, int ndim, const long *shape, const long *strides,
     int type_num, int itemsize, const std::shared_ptr<void> &base, int aflags,
     pybind11::return_value_policy policy, pybind11::handle parent);
