+++ include/gputils/system_utils.hpp
@@ -22,10 +22,10 @@ extern void mkdir_x(const std::string &path, int mode=0755);
 
 extern void mlockall_x(int flags = MCL_CURRENT | MCL_FUTURE | MCL_ONFAULT);
 
-extern void *mmap_x(void *addr, ssize_t length, int prot, int flags, int fd, off_t offset);
-extern void munmap_x(void *addr, ssize_t length);
+extern void *mmap_x(void *addr, long length, int prot, int flags, int fd, off_t offset);
+extern void munmap_x(void *addr, long length);
 
-extern void usleep_x(ssize_t usec);
+extern void usleep_x(long usec);
 
 
 // -------------------------------------------------------------------------------------------------
